adding folder for data processing, corresponding service
