--- twitter_bot/client/azure/__init__.py
@@ -1,3 +1,0 @@
-from .keyvault import KeyVaultClient
-from .storage import StorageClient
-from .ai_language import AILanguageClient
