+++ twitter_bot/data/__init__.py
@@ -1,0 +1,2 @@
+from .model import Politician
+from .services import PoliticianService
